+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * New: Add `get_filtered_color_scheme()` to get the filtered admin color scheme
 * New: Add support for constant to set admin color scheme, `C2C_FORCE_ADMIN_COLOR_SCHEME`
     * New: Add `is_constant_set()` to determine if constant was used to set the forced admin color scheme
+    * New: Add `get_color_scheme_via_constant()` to get the admin color scheme specified via the constant
     * New: Show message to users who can force an admin color scheme when the forced admin color scheme was set via the constant
 * New: Verify validity of admin color scheme on save and on retrieval
     * New: Add `is_valid_admin_color_scheme()` to check if a given admin color scheme is valid